pyproject.toml
[project]
name = "hangman-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi-users-db-sqlalchemy>=7.0.0",
    "fastapi-users[beanie]>=15.0.1",
    "fastapi[standard]>=0.121.1",
    "pwdlib>=0.2.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
